Reorganize godot-macros file structure
Organize by semantics, grouping class-related macros together.
Also move helper types into their own data_models directory.
