+++ godot-macros/src/class/mod.rs
@@ -1,0 +1,23 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+mod derive_godot_class;
+mod godot_api;
+mod data_models {
+    pub mod field;
+    pub mod field_export;
+    pub mod field_var;
+    pub mod func;
+    pub mod property;
+}
+
+pub(crate) use data_models::field::*;
+pub(crate) use data_models::field_export::*;
+pub(crate) use data_models::field_var::*;
+pub(crate) use data_models::func::*;
+pub(crate) use data_models::property::*;
+pub(crate) use derive_godot_class::*;
+pub(crate) use godot_api::*;
