+++ godot-macros/src/derive/mod.rs
@@ -1,0 +1,17 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+//! Derive macros on types outside of classes.
+
+mod derive_export;
+mod derive_from_variant;
+mod derive_property;
+mod derive_to_variant;
+
+pub(crate) use derive_export::*;
+pub(crate) use derive_from_variant::*;
+pub(crate) use derive_property::*;
+pub(crate) use derive_to_variant::*;
