+++ godot-macros/src/itest.rs
@@ -11,7 +11,7 @@ use venial::{Declaration, Error, FnParam, Function};
 use crate::util::{bail, path_ends_with, KvParser};
 use crate::ParseResult;
 
-pub fn transform(input_decl: Declaration) -> ParseResult<TokenStream> {
+pub fn attribute_itest(input_decl: Declaration) -> ParseResult<TokenStream> {
     let func = match input_decl {
         Declaration::Function(f) => f,
         _ => return bail!(&input_decl, "#[itest] can only be applied to functions"),
